chore: Implement prisma to create db schemas
This commit updates the npm dependencies in the `package.json` file to their latest stable versions. Specifically, it adds the `prisma` and `@prisma/client` dependencies with version `5.17.0`. This ensures that the project is using the most up-to-date packages and takes advantage of any bug fixes or new features introduced in these versions.